+++ db/migrate/20170809210610_create_stars.rb
@@ -1,7 +1,7 @@
 class CreateStars < ActiveRecord::Migration[5.1]
   def change
     create_table :stars do |t|
-      t.integer :mass
+      t.decimal :mass
 
       t.timestamps
     end
